+++ src/components/modules/index.ts
@@ -18,10 +18,8 @@ import UiAPI from './api/ui';
 
 /** ./toolbar */
 import BlockSettings from './toolbar/blockSettings';
-import ConversionToolbar from './toolbar/conversion';
 import Toolbar from './toolbar/index';
-// import InlineToolbar from './toolbar/inline';
-import InlineToolbar from './toolbar/inline2';
+import InlineToolbar from './toolbar/inline';
 
 /** . */
 import BlockEvents from './blockEvents';
@@ -60,7 +58,6 @@ export default {
 
   // Toolbar Modules
   BlockSettings,
-  ConversionToolbar,
   Toolbar,
   InlineToolbar,
