+++ src/components/modules/toolbar/inline.ts
@@ -3,30 +3,21 @@ import $ from '../../dom';
 import SelectionUtils from '../../selection';
 import * as _ from '../../utils';
 import { InlineTool as IInlineTool } from '../../../../types';
-import Flipper from '../../flipper';
 import I18n from '../../i18n';
 import { I18nInternalNS } from '../../i18n/namespace-internal';
 import Shortcuts from '../../utils/shortcuts';
-import * as tooltip from '../../utils/tooltip';
 import { ModuleConfig } from '../../../types-internal/module-config';
-import InlineTool from '../../tools/inline';
 import { CommonInternalSettings } from '../../tools/base';
-import { IconChevronDown } from '@codexteam/icons';
+import { Popover, PopoverItemParams, PopoverItemType } from '../../utils/popover';
+import { PopoverInline } from '../../utils/popover/popover-inline';
+import { getConvertToItems } from '../../utils/blocks';
+import { IconReplace } from '@codexteam/icons';
 
 /**
  * Inline Toolbar elements
  */
 interface InlineToolbarNodes {
   wrapper: HTMLElement | undefined;
-  togglerAndButtonsWrapper: HTMLElement | undefined;
-  buttons: HTMLElement | undefined;
-  conversionToggler: HTMLElement | undefined;
-  conversionTogglerContent: HTMLElement | undefined;
-  /**
-   * Zone below the buttons where Tools can create additional actions by 'renderActions()' method
-   * For example, input for the 'link' tool or textarea for the 'comment' tool
-   */
-  actions: HTMLElement | undefined;
 }
 
 /**
@@ -46,16 +37,6 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     inlineToolbarLeftOriented: 'ce-inline-toolbar--left-oriented',
     inlineToolbarRightOriented: 'ce-inline-toolbar--right-oriented',
     inlineToolbarShortcut: 'ce-inline-toolbar__shortcut',
-    buttonsWrapper: 'ce-inline-toolbar__buttons',
-    actionsWrapper: 'ce-inline-toolbar__actions',
-    inlineToolButton: 'ce-inline-tool',
-    inputField: 'cdx-input',
-    focusedButton: 'ce-inline-tool--focused',
-    conversionToggler: 'ce-inline-toolbar__dropdown',
-    conversionTogglerArrow: 'ce-inline-toolbar__dropdown-arrow',
-    conversionTogglerHidden: 'ce-inline-toolbar__dropdown--hidden',
-    conversionTogglerContent: 'ce-inline-toolbar__dropdown-content',
-    togglerAndButtonsWrapper: 'ce-inline-toolbar__toggler-and-button-wrapper',
   };
 
   /**
@@ -66,36 +47,25 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
   public opened = false;
 
   /**
+   * Popover instance reference
+   */
+  private popover: Popover | null = null;
+
+  /**
    * Margin above/below the Toolbar
    */
   // eslint-disable-next-line @typescript-eslint/no-magic-numbers
   private readonly toolbarVerticalMargin: number = _.isMobileScreen() ? 20 : 6;
 
   /**
-   * TODO: Get rid of this
-   *
    * Currently visible tools instances
    */
-  private toolsInstances: Map<string, IInlineTool>;
-
-  /**
-   * Buttons List
-   *
-   * @type {NodeList}
-   */
-  private buttonsList: NodeList = null;
+  private toolsInstances: Map<string, IInlineTool> = new Map();
 
   /**
    * Cache for Inline Toolbar width
-   *
-   * @type {number}
    */
   private width = 0;
-
-  /**
-   * Instance of class that responses for leafing buttons by arrows/tab
-   */
-  private flipper: Flipper = null;
 
   /**
    * @class
@@ -147,9 +117,8 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       return;
     }
 
-    await this.addToolsFiltered(needToShowConversionToolbar);
     this.move();
-    this.open(needToShowConversionToolbar);
+    await this.open(needToShowConversionToolbar);
     this.Editor.Toolbar.close();
   }
 
@@ -166,6 +135,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     }
 
     this.nodes.wrapper.classList.remove(this.CSS.inlineToolbarShowed);
+
     Array.from(this.toolsInstances.entries()).forEach(([name, toolInstance]) => {
       const shortcut = this.getToolShortcut(name);
 
@@ -184,8 +154,9 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     this.reset();
     this.opened = false;
 
-    this.flipper.deactivate();
-    this.Editor.ConversionToolbar.close();
+    this.popover?.hide();
+    this.popover?.destroy();
+    this.popover = null;
   }
 
   /**
@@ -205,15 +176,8 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    * Removes UI and its components
    */
   public destroy(): void {
-    /**
-     * Sometimes (in read-only mode) there is no Flipper
-     */
-    if (this.flipper) {
-      this.flipper.deactivate();
-      this.flipper = null;
-    }
-
     this.removeAllNodes();
+    this.popover?.destroy();
   }
 
   /**
@@ -230,48 +194,9 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     }
 
     /**
-     * Creates a different wrapper for toggler and buttons.
-     */
-    this.nodes.togglerAndButtonsWrapper = $.make('div', this.CSS.togglerAndButtonsWrapper);
-    this.nodes.buttons = $.make('div', this.CSS.buttonsWrapper);
-    this.nodes.actions = $.make('div', this.CSS.actionsWrapper);
-
-    // To prevent reset of a selection when click on the wrapper
-    this.listeners.on(this.nodes.wrapper, 'mousedown', (event) => {
-      const isClickedOnActionsWrapper = (event.target as Element).closest(`.${this.CSS.actionsWrapper}`);
-
-      // If click is on actions wrapper,
-      // do not prevent default behavior because actions might include interactive elements
-      if (!isClickedOnActionsWrapper) {
-        event.preventDefault();
-      }
-    });
-
-    /**
-     * Append the intermediary wrapper which contains toggler and buttons and button actions.
-     */
-    $.append(this.nodes.wrapper, [this.nodes.togglerAndButtonsWrapper, this.nodes.actions]);
-    /**
      * Append the inline toolbar to the editor.
      */
     $.append(this.Editor.UI.nodes.wrapper, this.nodes.wrapper);
-
-    /**
-     * Add button that will allow switching block type
-     */
-    this.addConversionToggler();
-
-    /**
-     * Wrapper for the inline tools
-     * Will be appended after the Conversion Toolbar toggler
-     */
-    $.append(this.nodes.togglerAndButtonsWrapper, this.nodes.buttons);
-
-    /**
-     * Prepare conversion toolbar.
-     * If it has any conversion tool then it will be enabled in the Inline Toolbar
-     */
-    this.prepareConversionToolbar();
 
     /**
      * Recalculate initial width with all buttons
@@ -280,18 +205,12 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     window.requestAnimationFrame(() => {
       this.recalculateWidth();
     });
-
-    /**
-     * Allow to leaf buttons by arrows / tab
-     * Buttons will be filled on opening
-     */
-    this.enableFlipper();
   }
 
   /**
    * Shows Inline Toolbar
    */
-  private open(): void {
+  private async open(): Promise<void> {
     if (this.opened) {
       return;
     }
@@ -301,18 +220,27 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
      */
     this.nodes.wrapper.classList.add(this.CSS.inlineToolbarShowed);
 
-    this.buttonsList = this.nodes.buttons.querySelectorAll(`.${this.CSS.inlineToolButton}`);
     this.opened = true;
 
-    /**
-     * Get currently visible buttons to pass it to the Flipper
-     */
-    let visibleTools = Array.from(this.buttonsList);
 
-    visibleTools.unshift(this.nodes.conversionToggler);
-    visibleTools = visibleTools.filter((tool) => !(tool as HTMLElement).hidden);
+    if (this.popover !== null) {
+      this.popover.destroy();
+    }
 
-    this.flipper.activate(visibleTools as HTMLElement[]);
+    const inlineTools = await this.getInlineTools();
+
+    this.popover = new PopoverInline({
+      items: inlineTools,
+      scopeElement: this.Editor.API.methods.ui.nodes.redactor,
+      messages: {
+        nothingFound: I18n.ui(I18nInternalNS.ui.popover, 'Nothing found'),
+        search: I18n.ui(I18nInternalNS.ui.popover, 'Filter'),
+      },
+    });
+
+    this.nodes.wrapper?.append(this.popover.getElement());
+
+    this.popover.show();
   }
 
   /**
@@ -410,90 +338,6 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
     this.width = this.nodes.wrapper.offsetWidth;
   }
 
-  /**
-   * Create a toggler for Conversion Dropdown
-   * and prepend it to the buttons list
-   */
-  private addConversionToggler(): void {
-    this.nodes.conversionToggler = $.make('div', this.CSS.conversionToggler);
-    this.nodes.conversionTogglerContent = $.make('div', this.CSS.conversionTogglerContent);
-
-    const iconWrapper = $.make('div', this.CSS.conversionTogglerArrow, {
-      innerHTML: IconChevronDown,
-    });
-
-    this.nodes.conversionToggler.appendChild(this.nodes.conversionTogglerContent);
-    this.nodes.conversionToggler.appendChild(iconWrapper);
-
-    this.nodes.togglerAndButtonsWrapper.appendChild(this.nodes.conversionToggler);
-
-    if (import.meta.env.MODE === 'test') {
-      this.nodes.conversionToggler.setAttribute('data-cy', 'conversion-toggler');
-    }
-
-    this.listeners.on(this.nodes.conversionToggler, 'click', () => {
-      this.Editor.ConversionToolbar.toggle((conversionToolbarOpened) => {
-        /**
-         * When ConversionToolbar is opening on activated InlineToolbar flipper
-         * Then we need to temporarily deactivate InlineToolbar flipper so that we could flip ConversionToolbar items
-         *
-         * Other case when ConversionToolbar is closing (for example, by escape) but we need to continue flipping
-         * InlineToolbar items, we activate InlineToolbar flipper
-         */
-        const canActivateInlineToolbarFlipper = !conversionToolbarOpened && this.opened;
-
-        if (canActivateInlineToolbarFlipper) {
-          this.flipper.activate();
-        } else if (this.opened) {
-          this.flipper.deactivate();
-        }
-      });
-    });
-
-    if (_.isMobileScreen() === false ) {
-      tooltip.onHover(this.nodes.conversionToggler, I18n.ui(I18nInternalNS.ui.inlineToolbar.converter, 'Convert to'), {
-        placement: 'top',
-        hidingDelay: 100,
-      });
-    }
-  }
-
-  /**
-   * Changes Conversion Dropdown content for current block's Tool
-   */
-  private async setConversionTogglerContent(): Promise<void> {
-    const { BlockManager } = this.Editor;
-    const { currentBlock } = BlockManager;
-    const toolName = currentBlock.name;
-
-    /**
-     * If tool does not provide 'export' rule, hide conversion dropdown
-     */
-    const conversionConfig = currentBlock.tool.conversionConfig;
-    const exportRuleDefined = conversionConfig && conversionConfig.export;
-
-    this.nodes.conversionToggler.hidden = !exportRuleDefined;
-    this.nodes.conversionToggler.classList.toggle(this.CSS.conversionTogglerHidden, !exportRuleDefined);
-
-    /**
-     * Get icon or title for dropdown
-     */
-    const toolboxSettings = await currentBlock.getActiveToolboxEntry() || {};
-
-    this.nodes.conversionTogglerContent.innerHTML =
-      toolboxSettings.icon ||
-      toolboxSettings.title ||
-      _.capitalize(toolName);
-  }
-
-  /**
-   * Makes the Conversion Dropdown
-   */
-  private prepareConversionToolbar(): void {
-    const ct = this.Editor.ConversionToolbar.make();
-
-    $.append(this.nodes.wrapper, ct);
-  }
 
   /**
    *  Working with Tools
@@ -501,106 +345,99 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    */
 
   /**
-   * Append only allowed Tools
-   *
-   * @param {boolean} needToShowConversionToolbar - pass false to not to show Conversion Toolbar (e.g. for Footnotes-like tools)
+   * Returns Inline Tools segregated by their appearance type: popover items and custom html elements.
+   * Sets this.toolsInstances map
    */
-  private async addToolsFiltered(needToShowConversionToolbar = true): Promise<void> {
+  private async getInlineTools(): Promise<PopoverItemParams[]> {
     const currentSelection = SelectionUtils.get();
     const currentBlock = this.Editor.BlockManager.getBlock(currentSelection.anchorNode as HTMLElement);
 
-    /**
-     * Clear buttons list
-     */
-    this.nodes.buttons.innerHTML = '';
-    this.nodes.actions.innerHTML = '';
-    this.toolsInstances = new Map();
+    const inlineTools = Array.from(currentBlock.tool.inlineTools.values());
 
-    Array.from(currentBlock.tool.inlineTools.values()).forEach(tool => {
-      this.addTool(tool);
-    });
+    const popoverItems = [] as PopoverItemParams[];
 
-    if (needToShowConversionToolbar && this.Editor.ConversionToolbar.hasTools()) {
-      /**
-       * Change Conversion Dropdown content for current tool
-       */
-      await this.setConversionTogglerContent();
-    } else {
-      /**
-       * hide Conversion Dropdown with there are no tools
-       */
-      this.nodes.conversionToggler.hidden = true;
-    }
+    /** Add "Convert to" */
+    const convertToItems = await getConvertToItems(currentBlock, this.Editor);
 
-    /**
-     * Recalculate width because some buttons can be hidden
-     */
-    this.recalculateWidth();
-  }
-
-  /**
-   * Add tool button and activate clicks
-   *
-   * @param {InlineTool} tool - InlineTool object
-   */
-  private addTool(tool: InlineTool): void {
-    const instance = tool.create();
-    const button = instance.render();
-
-    if (!button) {
-      _.log('Render method must return an instance of Node', 'warn', tool.name);
-
-      return;
-    }
-
-    button.dataset.tool = tool.name;
-    this.nodes.buttons.appendChild(button);
-    this.toolsInstances.set(tool.name, instance);
-
-    if (_.isFunction(instance.renderActions)) {
-      const actions = instance.renderActions();
-
-      this.nodes.actions.appendChild(actions);
-    }
-
-    this.listeners.on(button, 'click', (event) => {
-      this.toolClicked(instance);
-      event.preventDefault();
-    });
-
-    const shortcut = this.getToolShortcut(tool.name);
-
-    if (shortcut) {
-      try {
-        this.enableShortcuts(instance, shortcut);
-      } catch (e) {}
-    }
-
-    /**
-     * Enable tooltip module on button
-     */
-    const tooltipContent = $.make('div');
-    const toolTitle = I18n.t(
-      I18nInternalNS.toolNames,
-      tool.title || _.capitalize(tool.name)
-    );
-
-    tooltipContent.appendChild($.text(toolTitle));
-
-    if (shortcut) {
-      tooltipContent.appendChild($.make('div', this.CSS.inlineToolbarShortcut, {
-        textContent: _.beautifyShortcut(shortcut),
-      }));
-    }
-
-    if (_.isMobileScreen() === false ) {
-      tooltip.onHover(button, tooltipContent, {
-        placement: 'top',
-        hidingDelay: 100,
+    if (convertToItems.length > 0) {
+      popoverItems.push({
+        icon: IconReplace,
+        title: I18n.ui(I18nInternalNS.ui.popover, 'Convert to'),
+        children: {
+          searchable: true,
+          items: convertToItems,
+        },
+      });
+      popoverItems.push({
+        type: PopoverItemType.Separator,
       });
     }
 
-    instance.checkState(SelectionUtils.get());
+    inlineTools.forEach(tool => {
+      const instance = tool.create();
+      const controlData = instance.render();
+
+      this.toolsInstances.set(tool.name, instance);
+
+      /** Enable tool shortcut */
+      const shortcut = this.getToolShortcut(tool.name);
+
+      if (shortcut) {
+        try {
+          this.enableShortcuts(instance, shortcut);
+        } catch (e) {}
+      }
+
+      const shortcutBeautified = shortcut !== undefined ? _.beautifyShortcut(shortcut) : undefined;
+
+      const toolTitle = I18n.t(
+        I18nInternalNS.toolNames,
+        tool.title || _.capitalize(tool.name)
+      );
+
+      [ controlData ].flat().forEach((item) => {
+        let popoverItem = {
+          onActivate: () => {
+            this.toolClicked(instance);
+          },
+          hint: {
+            title: toolTitle,
+            description: shortcutBeautified,
+          },
+          isActive: instance.checkState(SelectionUtils.get()),
+        } as PopoverItemParams;
+
+        if ($.isElement(item)) {
+          popoverItem = {
+            ...popoverItem,
+            type: PopoverItemType.Html,
+            element: item,
+          };
+        } else {
+          popoverItem = {
+            ...popoverItem,
+            ...(item as PopoverItemParams),
+          };
+        }
+
+        if (_.isFunction(instance.renderActions)) {
+          const actions = instance.renderActions();
+
+          popoverItem.children = {
+            items: [
+              {
+                type: PopoverItemType.Html,
+                element: actions,
+              },
+            ],
+          };
+        }
+
+        popoverItems.push(popoverItem);
+      });
+    });
+
+    return popoverItems;
   }
 
   /**
@@ -608,7 +445,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
    *
    * @param toolName — Tool name
    */
-  private getToolShortcut(toolName): string | void {
+  private getToolShortcut(toolName: string): string | undefined {
     const { Tools } = this.Editor;
 
     /**
@@ -628,7 +465,7 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       return this.inlineTools[toolName][CommonInternalSettings.Shortcut];
     }
 
-    return tool.shortcut;
+    return tool?.shortcut;
   }
 
   /**
@@ -678,15 +515,6 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
 
     tool.surround(range);
     this.checkToolsState();
-
-    /**
-     * If tool has "actions", so after click it will probably toggle them on.
-     * For example, the Inline Link Tool will show the URL-input.
-     * So we disable the Flipper for that case to allow Tool bind own Enter listener
-     */
-    if (tool.renderActions !== undefined) {
-      this.flipper.deactivate();
-    }
   }
 
   /**
@@ -712,19 +540,5 @@ export default class InlineToolbar extends Module<InlineToolbarNodes> {
       });
 
     return result;
-  }
-
-  /**
-   * Allow to leaf buttons by arrows / tab
-   * Buttons will be filled on opening
-   */
-  private enableFlipper(): void {
-    this.flipper = new Flipper({
-      focusedItemClass: this.CSS.focusedButton,
-      allowedKeys: [
-        _.keyCodes.ENTER,
-        _.keyCodes.TAB,
-      ],
-    });
   }
 }
